+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/util/modifiable/MPRRange.java
@@ -6,7 +6,6 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonParseException;
 import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.annotations.SerializedName;
-import com.google.gson.reflect.TypeToken;
 import insane96mcp.insanelib.exception.JsonValidationException;
 import insane96mcp.mobspropertiesrandomness.data.json.IMPRObject;
 import insane96mcp.mobspropertiesrandomness.util.Logger;
@@ -113,7 +112,7 @@ public class MPRRange extends MPRModifiableValue implements IMPRObject {
 					context.deserialize(json.getAsJsonObject().get("world_spawn_distance_modifier"), MPRWorldSpawnDistanceModifier.class),
 					context.deserialize(json.getAsJsonObject().get("depth_modifier"), MPRDepthModifier.class),
 					context.deserialize(json.getAsJsonObject().get("time_existed_modifier"), MPRTimeExistedModifier.class),
-					context.deserialize(json.getAsJsonObject().get("conditions_modifier"), new TypeToken<List<MPRConditionModifier>>() {}.getType()),
+					context.deserialize(json.getAsJsonObject().get("condition_modifiers"), MPRConditionModifier.LIST_TYPE),
 					context.deserialize(json.getAsJsonObject().get("round"), Integer.class));
 		}
 	}