+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added `team` to properties to add an entity to a scoreboard team
 * Fixed some NBT tags not getting applied
+* Fixed `condition_modifiers` not working
 
 ## 4.10.8
 * Fixed boss bar ignoring conditions